+++ .env
@@ -1,0 +1,2 @@
+# BDD ENvironment Variables
+DATABASE_URL = "postgresql://user:user@localhost:5432/postgres"
