+++ pyproject.toml
@@ -1,31 +1,49 @@
 [project]
-name = "trackntrain"
+name = "tracknatrainapi"
 version = "0.1.0"
-description = ""
-authors = [
-    {name = "Baptiste-Ferrand",email = "baptiste.ferrand@ynov.com"}
-]
-readme = "README.md"
-requires-python = ">=3.10"
+requires-python = ">=3.13"
 dependencies = [
-    "fastapi (>=0.115.13,<0.116.0)",
-    "uvicorn[standard] (>=0.34.3,<0.35.0)"
+    "annotated-types==0.7.0",
+    "anyio==4.9.0",
+    "bcrypt==4.3.0",
+    "boto3==1.37.37",
+    "botocore==1.37.37",
+    "cffi==1.17.1",
+    "click==8.1.8",
+    "cryptography==44.0.2",
+    "dnspython==2.7.0",
+    "ecdsa==0.19.1",
+    "email-validator==2.2.0",
+    "exceptiongroup==1.2.2",
+    "fastapi==0.115.12",
+    "greenlet==3.1.1",
+    "h11==0.14.0",
+    "idna==3.10",
+    "jmespath==1.0.1",
+    "psycopg2-binary==2.9.10",
+    "pyasn1==0.4.8",
+    "pycparser==2.22",
+    "pydantic==2.11.3",
+    "pydantic-core==2.33.1",
+    "python-dateutil==2.9.0.post0",
+    "python-dotenv==1.1.0",
+    "python-jose==3.4.0",
+    "python-multipart==0.0.20",
+    "rsa==4.9",
+    "s3transfer==0.11.5",
+    "six==1.17.0",
+    "sniffio==1.3.1",
+    "sqlalchemy==2.0.40",
+    "starlette==0.46.2",
+    "typing-extensions==4.13.2",
+    "typing-inspection==0.4.0",
+    "urllib3==2.4.0",
+    "uvicorn==0.34.1",
 ]
-
 
 [build-system]
-requires = ["poetry-core>=2.0.0,<3.0.0"]
-build-backend = "poetry.core.masonry.api"
+requires = ["setuptools>=42", "wheel"]
+build-backend = "setuptools.build_meta"
 
-[tool.poetry.group.dev.dependencies]
-pytest = "^8.4.1"
-black = "^25.1.0"
-isort = "^6.0.1"
-mypy = "^1.16.1"
-alembic = "^1.16.2"
-
-[tool.poetry.scripts]
-start-api = "uvicorn app.entrypoints.api.main:app --reload"
-
-[tool.poetry]
-package-mode = false
+[tool.setuptools.packages.find]
+include = ["app*", "alembic*"]
