+++ src/domain/exceptions.py
@@ -1,0 +1,7 @@
+class DomainError(Exception):
+    """Racine pour toutes les erreurs métier."""
+    pass
+
+class DuplicateProfileError(DomainError):
+    """Email déjà utilisé."""
+    pass
