+++ src/domain/ports/password_hasher.py
@@ -1,0 +1,10 @@
+from abc import ABC, abstractmethod
+
+class PasswordHasher(ABC):
+    @abstractmethod
+    def hash(self, plain: str) -> str:
+        pass
+
+    @abstractmethod
+    def verify(self, plain: str, hashed: str) -> bool:
+        pass
